changed seed from ulong to int

+++ Scripts/World/SaveScripts/WorldData.cs
@@ -8,9 +8,9 @@ public class WorldData
 
     public List<WorldPos> columns = new List<WorldPos>();
     public string worldName;
-    public ulong worldSeed;
+    public int worldSeed;
 
-    public WorldData(string name, ulong seed){
+    public WorldData(string name, int seed){
         worldName = name;
         worldSeed = seed;
     }
